api/requirements.txt
fastapi == 0.111.0
langchain
langchain-openai
azure-cognitiveservices-speech == 1.45.0
python-multipart == 0.0.9
langchain-community
azure-search-documents == 11.4.0
azure-identity == 1.17.1
azure-communication-callautomation == 1.1.0
docx == 0.2.4
python-docx == 1.1.2
docx2txt == 0.8
langdetect == 1.0.9
pyngrok
azure-ai-projects
dotenv
Quart>=0.19.6
azure-eventgrid==4.11.0
azure-communication-callautomation==1.1.0
openai>=1.0.0
